Wire up DisplayImage
