+++ .sbt_metadata/ingestor_images.json
@@ -2,6 +2,7 @@
   "id" : "ingestor_images",
   "folder" : "pipeline/ingestor/ingestor_images",
   "dependencyIds" : [
-    "ingestor_common"
+    "ingestor_common",
+    "display_model"
   ]
 }
